Share source control presentation helpers
